+++ projects/network/network-core/src/commonTest/kotlin/ktor/serverError/Error505Test.kt
@@ -3,7 +3,6 @@ package com.javiersc.either.network.ktor.serverError
 import com.javiersc.either.network.NetworkEither
 import com.javiersc.either.network.buildNetworkFailureHttp
 import com.javiersc.either.network.ktor.BaseTest
-import com.javiersc.either.network.ktor.NetworkEither
 import com.javiersc.either.network.models.DogDTO
 import com.javiersc.either.network.models.ErrorDTO
 import com.javiersc.either.network.runTestBlocking
@@ -19,6 +18,6 @@ internal class Error505Test : BaseTest<ErrorDTO, DogDTO>() {
 
     @Test
     fun `Request 505`() = runTestBlocking {
-        NetworkEither<ErrorDTO, DogDTO> { client.get("path") } shouldBe expected
+        networkEitherKtor<ErrorDTO, DogDTO> { get("path") } shouldBe expected
     }
 }