+++ projects/network/network-core/src/commonTest/kotlin/ktor/success/Success204Test.kt
@@ -3,7 +3,6 @@ package com.javiersc.either.network.ktor.success
 import com.javiersc.either.network.NetworkEither
 import com.javiersc.either.network.buildNetworkSuccess
 import com.javiersc.either.network.ktor.BaseNullTest
-import com.javiersc.either.network.ktor.NetworkEither
 import com.javiersc.either.network.models.ErrorDTO
 import com.javiersc.either.network.runTestBlocking
 import io.kotest.matchers.shouldBe
@@ -17,6 +16,6 @@ internal class Success204Test : BaseNullTest<ErrorDTO>() {
 
     @Test
     fun `Request 204`() = runTestBlocking {
-        NetworkEither<ErrorDTO, Unit> { client.get("path") } shouldBe expected
+        networkEitherKtor<ErrorDTO, Unit> { get("path") } shouldBe expected
     }
 }